refactor(RacingGame): extract game logic from main into RacingGame object - moved game logic from main to RacingGame for better readability and structure
